% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkdataset}
\alias{checkdataset}
\title{Checkdataset}
\usage{
checkdataset(year, pollutant = NULL, lis = NULL)
}
\arguments{
\item{year}{character. The year of the dataset to check}

\item{pollutant}{character. The name of the pollutant to check. If \code{NULL} (the default) means we are checking the dataset for the stations}

\item{lis}{list. The list contains on the first slot the dataset requested (using \code{pollutant} parameter) and on the second slot uploaded the hash to maintain the dataset in memoery.}
}
\value{
data.frame
}
\description{
Check whether and which dataset is already loaded, download and visualise it.
If the dataset is loaded, then get that from an \code{hash}, if not loaded then call \code{\link{restfullAPI}} and upload the \code{hash}
}
\examples{
checkdataset(year, pollutant, lis)



}
