% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{datacleaning}
\alias{datacleaning}
\title{Clean dataset TS}
\usage{
datacleaning(leggo)
}
\arguments{
\item{leggo}{data.frame. Dataframe to clean}
}
\value{
data.frame. Dataframe cleaned
}
\description{
Clean the dataset for time series. It will delete the station_id, remove NA and convert the classes of the objects.
Then, aggregate the value detected in the same day using a mean function
}
\examples{
\dontrun{
datacleaning(leggo)
}
}
