\name{installpack}
\alias{installpack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Install missing packages}
\description{
Install every missing packages needed by the application milanpollution.
}
\usage{
installpack()
}

\author{
Andrea Ierardi
}


\examples{

installpack()

}

