% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restfullAPI.R
\name{restfullAPI}
\alias{restfullAPI}
\title{restfullAPI}
\usage{
restfullAPI(id = NULL)
}
\arguments{
\item{id}{character. Identify the id of the dataset. if \code{NULL} return an error}
}
\value{
data.frame
}
\description{
Get the dataset from the Opendata website of the city of Milan using CRAN API.
}
\examples{
\dontrun{
restfullAPI(id)
}
}
