% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{station_clean}
\alias{station_clean}
\title{Clean dataset Stations}
\usage{
station_clean(file)
}
\arguments{
\item{file}{data.frame. Dataframe to clean}
}
\value{
data.frame. Dataframe cleaned
}
\description{
Clean the dataset for station informations. Convert the classes of the object, remove NA and create a new dataframe with the station informations.
}
\examples{
\dontrun{
station_clean(file)
}
}
