% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkdataset}
\alias{checkdataset}
\title{Check dataset to upload}
\usage{
checkdataset(year, pollutant)
}
\arguments{
\item{pollutant}{}
}
\value{

}
\description{
Check dataset to upload
}
\examples{
\dontrun{
checkdataset(year, pollutant)
}


}
