% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkdataset}
\alias{checkdataset}
\title{Checkdataset}
\usage{
checkdataset(year, pollutant)
}
\arguments{
\item{year}{character. The year of the dataset to check}

\item{pollutant}{character. The name of the pollutant to check. if \code{NULL} (the default) means we are checking the dataset for the stations}
}
\value{
data.frame
}
\description{
Check which and whether the dataset is already loaded, download and visualise it.
If the dataset is loaded, then get that from an \code{hash}, if not loaded then call \code{\link{restfullAPI}} and upload the \code{hash}
}
\examples{
\dontrun{
checkdataset(year, pollutant)
}


}
