% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkdataset}
\alias{checkdataset}
\title{Checkdataset}
\usage{
checkdataset(year, pollutant)
}
\arguments{
\item{year}{character. The year of the dataset to check}

\item{pollutant}{character. The name of the pollutant to check. if \code{NULL} (the default) means we are checking the dataset for the stations}
}
\value{
data.frame
}
\description{
Check which dataset download and visualise
}
\examples{
\dontrun{
checkdataset(year, pollutant)
}


}
