% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkdataset}
\alias{checkdataset}
\title{Checkdataset}
\usage{
checkdataset(year, pollutant = NULL, lis = NULL)
}
\arguments{
\item{year}{character. The year of the dataset to check}

\item{pollutant}{character. The name of the pollutant to check. If \code{NULL} (the default) means we are checking the dataset for the stations else we are checking for time series.}

\item{lis}{list. The list contains on the first slot the dataset requested (using \code{pollutant} parameter) and on the second slot uploaded the hash to maintain the dataset in memory.}
}
\value{
list
}
\description{
Check whether and which datasetis already loaded, download and visualise it. \cr
If the dataset is loaded, then get that from an \code{hash}, if not loaded then call \link{restfullAPI}
and upload the \code{hash} If pollutant is \code{NULL} then the function will get the time series dataset
else the function will get the data detected from each station. It return a list contains the dataset cleaned
and requested and a list containing the hash of the dataset already download to keep it in memory.
\cr PS. This only works with "Rilevazione qualita' aria" dataset of openData website!
}
\examples{
\dontrun{
checkdataset(year, pollutant, lis)
}
}
