% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset.R
\name{checkYears}
\alias{checkYears}
\title{Check years Milan Pollution}
\usage{
checkYears(year, lis)
}
\arguments{
\item{year}{character. The year of the dataset to check}

\item{lis}{list. List that contains the hash of saved dataset}
}
\value{
list
}
\description{
Check if dataset of years 2017, 2018 or 2019 is already loaded
}
\examples{
\dontrun{
checkYears(year, lis)
}
}
