\name{installpack}
\alias{installpack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Install missing packages}
\description{
Install every missing packages needed by the application milanpollution. 
The function will install all these libraries : shiny, ggplot2, forecast, xts, ckanr, httr, jsonlite, tidyverse and plotly.
}
\usage{
installpack()
}

\author{
Andrea Ierardi
}


\examples{

installpack()

}

