% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restfullAPI.R
\name{restfullAPI}
\alias{restfullAPI}
\title{restfullAPI}
\usage{
restfullAPI(id = NULL)
}
\arguments{
\item{id}{character. Identify the ID of the dataset. if \code{NULL} or not a valid ID for a dataset, return an error.}
}
\value{
data.frame
}
\description{
Get the dataset from the Opendata website of the city of Milan using CRAN API.
}
\examples{
restfullAPI(id)

}
